Atari env dev

See merge request open-XLab/cell/nerveX!11

+++ nervex/envs/gym/cartpole/action/cartpole_action.py
@@ -15,7 +15,7 @@ class CartpoleRawAction(EnvElement):
     def _init(self):
         self._default_val = None
         self._shape = (1, )
-        self._value = {'min': 0, 'max': 1, 'dtype': int, 'dinfo': 'discreate int value, 0 is go left and 1 is go right'}
+        self._value = {'min': 0, 'max': 1, 'dtype': torch.LongTensor, 'dinfo': 'discreate int value, 0 is go left and 1 is go right'}
 
     def _to_agent_processor(self, action):
         return action
